Disable selinux

+++ scripts/install.sh
@@ -178,9 +178,18 @@ assert_installed(){
 }
 
 
+#
+
+
+
+
 
 assert_keitaro_not_installed(){
   debug 'Ensure keitaro is not installed yet'
+  if isset "$RECONFIGURE"; then
+    debug 'Skip checking install.lock'
+    return
+  fi
   if is_exists_file /var/www/keitaro/var/install.lock no; then
     debug 'NOK: keitaro is already installed'
     fail "$(translate errors.keitaro_already_installed)"
@@ -1999,7 +2008,9 @@ json2dict() {
 
 
 write_emtpy_hosts_txt(){
-  echo -e "[server]\nlocalhost connection=local ansible_user=root" > hosts.txt
+  if [[ ! -f hosts.txt ]]; then
+    echo -e "[server]\nlocalhost connection=local ansible_user=root" > hosts.txt
+  fi
 }
 
 install(){
